# webapp/contracts/openapi/bundled.yaml
openapi: 3.1.0
info:
  title: BizSim Edu API
  description: 商域 - 商业模拟教育平台后端 API
  version: 1.0.0
paths:
...
  /api/v1/competitions:
    get:
      tags:
      - 比赛
      summary: List Competitions
      description: 获取公开比赛列表（默认仅正式赛）
      operationId: list_competitions_api_v1_competitions_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: match_kind
        in: query
        required: false
        schema:
          type: string
          default: official
          title: Match Kind
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_List_CompetitionEventOut__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - 比赛
      summary: Create Competition
      description: 创建比赛
      operationId: create_competition_api_v1_competitions_post
      security:
      - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompetitionEventCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CompetitionEventOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/competitions/{event_id}:
    get:
      tags:
      - 比赛
      summary: Get Competition
      description: 获取比赛详情
      operationId: get_competition_api_v1_competitions__event_id__get
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
          title: Event Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CompetitionEventOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - 比赛
      summary: Update Competition
      description: 更新比赛信息
      operationId: update_competition_api_v1_competitions__event_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
          title: Event Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompetitionEventUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CompetitionEventOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/competitions/join:
    post:
      tags:
      - 比赛
      summary: Join Competition
      description: 通过房间码加入比赛
      operationId: join_competition_api_v1_competitions_join_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinCompetitionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ParticipantOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/competitions/{event_id}/leave:
    post:
      tags:
      - 比赛
      summary: Leave Competition
      description: 退出比赛
      operationId: leave_competition_api_v1_competitions__event_id__leave_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
          title: Event Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_dict_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/competitions/{event_id}/standings:
    get:
      tags:
      - 比赛
      summary: Get Standings
      description: 获取排行榜
      operationId: get_standings_api_v1_competitions__event_id__standings_get
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
          title: Event Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_List_dict__'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/competitions/{event_id}/my-status:
    get:
      tags:
      - 比赛
      summary: Get My Status
      description: 获取我的参赛状态
      operationId: get_my_status_api_v1_competitions__event_id__my_status_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
          title: Event Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_MyCompetitionStatus_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/competitions/{event_id}/start:
    post:
      tags:
      - 比赛
      summary: Start Competition
      description: 开始比赛
      operationId: start_competition_api_v1_competitions__event_id__start_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
          title: Event Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CompetitionEventOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/competitions/{event_id}/end:
    post:
      tags:
      - 比赛
      summary: End Competition
      description: 结束比赛并计算结果
      operationId: end_competition_api_v1_competitions__event_id__end_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
          title: Event Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CompetitionEventOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiResponse_CompetitionEventOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        code:
          type: integer
          title: Code
          default: 0
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/CompetitionEventOut'
          - type: 'null'
      type: object
      title: ApiResponse[CompetitionEventOut]
...
    ApiResponse_List_CompetitionEventOut__:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        code:
          type: integer
          title: Code
          default: 0
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/CompetitionEventOut'
            type: array
          - type: 'null'
          title: Data
      type: object
      title: ApiResponse[List[CompetitionEventOut]]
...
    ApiResponse_List_dict__:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        code:
          type: integer
          title: Code
          default: 0
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Data
      type: object
      title: ApiResponse[List[dict]]
...
    ApiResponse_MyCompetitionStatus_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        code:
          type: integer
          title: Code
          default: 0
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/MyCompetitionStatus'
          - type: 'null'
      type: object
      title: ApiResponse[MyCompetitionStatus]
...
    ApiResponse_ParticipantOut_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        code:
          type: integer
          title: Code
          default: 0
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - $ref: '#/components/schemas/ParticipantOut'
          - type: 'null'
      type: object
      title: ApiResponse[ParticipantOut]
...
    ApiResponse_dict_:
      properties:
        success:
          type: boolean
          title: Success
          default: true
        code:
          type: integer
          title: Code
          default: 0
        message:
          type: string
          title: Message
          default: success
        data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Data
      type: object
      title: ApiResponse[dict]
...
    CompetitionEventCreate:
      properties:
        title:
          type: string
          maxLength: 100
          minLength: 1
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        game_type:
          type: string
          title: Game Type
          default: trading
        max_players:
          type: integer
          maximum: 200.0
          minimum: 2.0
          title: Max Players
          default: 50
        config:
          $ref: '#/components/schemas/GameConfig'
        game_config_id:
          type: string
          maxLength: 64
          title: Game Config Id
          default: trading-v2-rts
        design_mode:
          type: string
          title: Design Mode
          description: standalone | modular
          default: standalone
        match_kind:
          type: string
          title: Match Kind
          description: official | practice
          default: official
        teaching_group_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Teaching Group Id
          description: 所属体验营营团 ID
      type: object
      required:
      - title
      title: CompetitionEventCreate
    CompetitionEventOut:
      properties:
        id:
          type: integer
          title: Id
        organizer_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Organizer Id
        teaching_group_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Teaching Group Id
        room_code:
          type: string
          title: Room Code
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        match_kind:
          type: string
          title: Match Kind
          default: official
        design_mode:
          type: string
          title: Design Mode
          default: standalone
        game_config_id:
          type: string
          title: Game Config Id
          default: trading-v1
        game_type:
          type: string
          title: Game Type
        status:
          type: string
          title: Status
        config:
          additionalProperties: true
          type: object
          title: Config
        max_players:
          type: integer
          title: Max Players
        current_round:
          type: integer
          title: Current Round
        starts_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Starts At
        ends_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Ends At
        created_at:
          type: string
          format: date-time
          title: Created At
        participant_count:
          type: integer
          title: Participant Count
          default: 0
      type: object
      required:
      - id
      - room_code
      - title
      - description
      - game_type
      - status
      - config
      - max_players
      - current_round
      - starts_at
      - ends_at
      - created_at
      title: CompetitionEventOut
    CompetitionEventUpdate:
      properties:
        title:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      type: object
      title: CompetitionEventUpdate
...
    GameConfig:
      properties:
        rounds:
          type: integer
          maximum: 20.0
          minimum: 3.0
          title: Rounds
          default: 10
        initial_capital:
          type: integer
          maximum: 100000.0
          minimum: 10000.0
          title: Initial Capital
          default: 50000
        inventory_limit:
          type: integer
          maximum: 999.0
          minimum: 5.0
          title: Inventory Limit
          description: v1：每种商品上限
          default: 99
        inventory_limit_per_product:
          type: integer
          maximum: 999.0
          minimum: 5.0
          title: Inventory Limit Per Product
          default: 99
        move_cost:
          type: integer
          maximum: 5000.0
          minimum: 100.0
          title: Move Cost
          default: 1000
        decision_time:
          type: integer
          maximum: 300.0
          minimum: 15.0
          title: Decision Time
          default: 60
        mode:
          anyOf:
          - type: string
          - type: 'null'
          title: Mode
          description: rts 即时战略
        duration_preset:
          anyOf:
          - type: string
          - type: 'null'
          title: Duration Preset
          description: short | standard | long
        cities:
          items:
            type: string
          type: array
          title: Cities
          description: 默认长三角六城；RTS 赛制由 world.region_id 注入
          default:
          - nanjing
          - suzhou
          - shanghai
          - nantong
          - wuxi
          - changzhou
        products:
          items:
            type: string
          type: array
          title: Products
          default:
          - fruit
          - vegetable
          - daily
          - electronics
          - clothing
          - cosmetics
          - jewelry
          - antique
          - art
          - snack
      additionalProperties: true
      type: object
      title: GameConfig
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    JoinCompetitionRequest:
      properties:
        room_code:
          type: string
          maxLength: 4
          minLength: 4
          pattern: ^\d{4}$
          title: Room Code
      type: object
      required:
      - room_code
      title: JoinCompetitionRequest
...
    MyCompetitionStatus:
      properties:
        event:
          $ref: '#/components/schemas/CompetitionEventOut'
        participant:
          anyOf:
          - $ref: '#/components/schemas/ParticipantOut'
          - type: 'null'
        is_organizer:
          type: boolean
          title: Is Organizer
      type: object
      required:
      - event
      - participant
      - is_organizer
      title: MyCompetitionStatus
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login